# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HackYeah2026 API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cities": {
      "get": {
        "tags": [
          "cities"
        ],
        "summary": "List Cities",
        "operationId": "list_cities_api_cities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CityConfig"
                  },
                  "type": "array",
                  "title": "Response List Cities Api Cities Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/places": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "List Places",
        "operationId": "list_places_api_places_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "krakow",
              "title": "City"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Szukaj po nazwie",
              "title": "Q"
            },
            "description": "Szukaj po nazwie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Place"
                  },
                  "title": "Response List Places Api Places Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/routes": {
      "post": {
        "tags": [
          "routes"
        ],
        "summary": "Plan Route",
        "operationId": "plan_route_api_routes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Create Report",
        "operationId": "create_report_api_reports_post",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "krakow",
              "title": "City"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Reports",
        "operationId": "list_reports_api_reports_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "krakow",
              "title": "City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Report"
                  },
                  "title": "Response List Reports Api Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccessibilityAttribute": {
        "properties": {
          "key": {
            "$ref": "#/components/schemas/AttributeKey"
          },
          "value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "0 = nie wiemy, 1 = pewne"
          },
          "status": {
            "$ref": "#/components/schemas/AttributeStatus",
            "default": "unverified"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/AccessibilityAttribute"
            },
            "type": "array",
            "title": "Alternatives"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "provenance",
          "confidence"
        ],
        "title": "AccessibilityAttribute"
      },
      "AttributeKey": {
        "type": "string",
        "enum": [
          "wheelchair",
          "step_free_entrance",
          "stairs",
          "step_count",
          "ramp",
          "elevator",
          "accessible_toilet",
          "surface",
          "incline_percent",
          "kerb_height_cm",
          "width_cm",
          "tactile_paving"
        ],
        "title": "AttributeKey",
        "description": "Słownik cech dostępności. Dodajemy nowe klucze tutaj, nie ad hoc w kodzie."
      },
      "AttributeStatus": {
        "type": "string",
        "enum": [
          "verified",
          "unverified",
          "conflicting",
          "outdated"
        ],
        "title": "AttributeStatus"
      },
      "CityConfig": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "bbox": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox"
          },
          "center": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Center"
          },
          "default_zoom": {
            "type": "integer",
            "title": "Default Zoom",
            "default": 13
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderConfig"
            },
            "type": "array",
            "title": "Providers"
          },
          "routing": {
            "$ref": "#/components/schemas/RoutingConfig"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "country",
          "timezone",
          "bbox",
          "center"
        ],
        "title": "CityConfig"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "database": {
            "type": "string",
            "title": "Database"
          }
        },
        "type": "object",
        "required": [
          "status",
          "database"
        ],
        "title": "HealthResponse"
      },
      "LatLon": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "LatLon"
      },
      "Place": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stabilne ID, np. 'osm:node/123'"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "np. 'cafe', 'pharmacy', 'tram_stop'"
          },
          "location": {
            "$ref": "#/components/schemas/LatLon"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/AccessibilityAttribute"
            },
            "type": "array",
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city",
          "location"
        ],
        "title": "Place"
      },
      "Provenance": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Nazwa providera, np. 'osm', 'msip'"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref",
            "description": "ID/URL rekordu w źródle, np. 'node/123456'"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "last_verified": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified",
            "description": "Kiedy informacja była ostatnio potwierdzona w źródle"
          }
        },
        "type": "object",
        "required": [
          "source",
          "source_type",
          "fetched_at"
        ],
        "title": "Provenance"
      },
      "ProviderConfig": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProviderConfig"
      },
      "Report": {
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LatLon"
          },
          "place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          },
          "attribute": {
            "$ref": "#/components/schemas/AttributeKey"
          },
          "value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "status": {
            "$ref": "#/components/schemas/ReportStatus",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "location",
          "attribute",
          "value",
          "id",
          "city",
          "created_at"
        ],
        "title": "Report"
      },
      "ReportCreate": {
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LatLon"
          },
          "place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          },
          "attribute": {
            "$ref": "#/components/schemas/AttributeKey"
          },
          "value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "location",
          "attribute",
          "value"
        ],
        "title": "ReportCreate",
        "description": "Zgłoszenie użytkownika. Celowo bez danych osobowych (patrz docs/security-privacy.md)."
      },
      "ReportStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed",
          "rejected",
          "resolved"
        ],
        "title": "ReportStatus"
      },
      "RoutePreferences": {
        "properties": {
          "avoid_stairs": {
            "type": "boolean",
            "title": "Avoid Stairs",
            "default": true
          },
          "max_incline_percent": {
            "type": "number",
            "maximum": 30.0,
            "minimum": 0.0,
            "title": "Max Incline Percent",
            "default": 6.0
          },
          "max_kerb_height_cm": {
            "type": "number",
            "maximum": 30.0,
            "minimum": 0.0,
            "title": "Max Kerb Height Cm",
            "default": 3.0
          },
          "avoid_rough_surface": {
            "type": "boolean",
            "title": "Avoid Rough Surface",
            "default": true
          },
          "prefer_lit_paths": {
            "type": "boolean",
            "title": "Prefer Lit Paths",
            "default": false
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile",
            "description": "Gotowy preset: 'wheelchair' lub 'stroller' (opcjonalnie)"
          }
        },
        "type": "object",
        "title": "RoutePreferences",
        "description": "Preferencje trasy. Pytamy o potrzeby, NIE o niepełnosprawność."
      },
      "RouteRequest": {
        "properties": {
          "city": {
            "type": "string",
            "title": "City",
            "default": "krakow"
          },
          "origin": {
            "$ref": "#/components/schemas/LatLon"
          },
          "destination": {
            "$ref": "#/components/schemas/LatLon"
          },
          "preferences": {
            "$ref": "#/components/schemas/RoutePreferences"
          }
        },
        "type": "object",
        "required": [
          "origin",
          "destination"
        ],
        "title": "RouteRequest"
      },
      "RouteResponse": {
        "properties": {
          "distance_m": {
            "type": "number",
            "title": "Distance M"
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/RouteSegment"
            },
            "type": "array",
            "title": "Segments"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "distance_m",
          "duration_s",
          "segments"
        ],
        "title": "RouteResponse"
      },
      "RouteSegment": {
        "properties": {
          "instruction": {
            "type": "string",
            "title": "Instruction",
            "description": "Np. 'Skręć w prawo w ul. Floriańską'"
          },
          "distance_m": {
            "type": "number",
            "title": "Distance M"
          },
          "geometry": {
            "items": {
              "$ref": "#/components/schemas/LatLon"
            },
            "type": "array",
            "title": "Geometry"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "incline_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incline Percent"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "data_status": {
            "$ref": "#/components/schemas/AttributeStatus",
            "default": "unverified"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "instruction",
          "distance_m",
          "geometry",
          "confidence"
        ],
        "title": "RouteSegment",
        "description": "Odcinek trasy z opisem tekstowym - to jest tekstowa alternatywa mapy (WCAG)."
      },
      "RoutingConfig": {
        "properties": {
          "network_type": {
            "type": "string",
            "title": "Network Type",
            "default": "walk"
          }
        },
        "type": "object",
        "title": "RoutingConfig"
      },
      "SourceType": {
        "type": "string",
        "enum": [
          "osm",
          "open_data",
          "msip",
          "user_report",
          "manual"
        ],
        "title": "SourceType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}